fix: move dimension-axis mapping into package module
